app-python/requirements.txt
django
mysqlclient
pyjwt
django-redis
djangorestframework
bcrypt
ebooklib
WeasyPrint
pycairo
pillow
weasyprint
